requirements.txt
python-dateutil~=2.8.2
six==1.16.0
